requirements.txt
ultralytics
streamlit
pillow
torch
torchvision
opencv-python
pandas
plotly
requests
matplotlib
numpy
